Asintel 2749 v2 (#386)
* added the max column

* master to branch (#383)

* Added Pipeline Controller

* Updated default for last updated timestamp

* Fix

* added SP

* Updated sql proj

* Added Controller Metadata

* Added PopulatePipelineController To Build


* master to branch (#384)

* Added Pipeline Controller

* Updated default for last updated timestamp

* Fix

* added SP

* Updated sql proj

* Added Controller Metadata

* Added PopulatePipelineController To Build


* updated the table and stored procedure

* adding county and  [Va_LocalAuthority]

* added the  county and local Authority to new stored procedure

* updated

* removed the fb prefiex and tried hence working

+++ src/SFA.DAS.DATA.MANAGEMENT/StoredProcedure/ImportVacancies.sql
@@ -387,7 +387,7 @@ INSERT INTO [ASData_PL].[Va_Vacancy]
            ,[SourceVacancyId]
            ,[SourceDb])
    SELECT  cast(v.BinaryId as varchar(256))                        as VacancyGuid
-	      ,cast(VacancyReference as int)                           as VacancyReference
+	      ,cast(VacancyReference as bigint)                           as VacancyReference
 		  ,cast(VacancyStatus as varchar(100))                     as VacancyStatus
 		  ,VacancyTitle                                            as VacancyTitle
 		  ,CASE WHEN len(EmployerPostCode)>8
